Added padding to PIN layout status and carrier text
When using a PIN backup method, Face Unlock was touching the bottom of
the status text on the top and the top of the carrier text on the
bottom, which made it look crowded and bad.

4 dips of padding was added to the bottom of the status text and
another 4 dips was added to the top of the carrier text.  This
corresponds to 8 pixels on the prime and 6 pixels on the crespo.  The
effect of this is that the keypad is 6 pixles closer to the entry text
on the crespo and 8 pixels closer to the entry text on the prime.  On
crespo, the keypad and entry text views are now as close as possible
when examining the pixel locations, but visually there is still space
between them and the difference is hardly noticable to the eye.

Despite the small change, Face Unlock looks a lot better when it is
not literally touching the text.

Change-Id: I01a7a6ace4789232595c5aa7683c1342ebbb2ee2