Allow override of p2p dialogs
By default, we show dialogs for connection requests and for
user interaction for pin input from the framework.

For applications, that want to make the experience more seamless
we should allow automatic discovery and connection as long
as one app can control it. We allow this for a foreground
app alone right now.

This will be a hidden API for now.

Bug: 6411069
Change-Id: Id342e933073d30eb58bf5a03a47ca26a64df8ddb
Signed-off-by: isheriff@google.com
Signed-off-by: Yoshihiko Ikenaga <yoshihiko.ikenaga@jp.sony.com>