Check in older versions of RS tests.
This change adds a v11 version of Fountain, as well as v11 and v14 versions of
RSTest. All tests currently build and pass on the required platforms.

This also fixes an initialization bug for RSTest running on older devices.
If we don't directly initialize "each" component of the vector types, it can
result in uninitialized memory (or zero) being used for divisors (particularly
with the char vector types). This triggers SIGFPE with the latest pending
Clang/LLVM rebase.

Change-Id: I92a70c2aeed4e20f110b1cb24ee926fc27be9dbb